+++ src/main/java/com/gamemoonchul/application/converter/PostConverter.java
@@ -1,8 +1,16 @@
 package com.gamemoonchul.application.converter;
 
+import com.gamemoonchul.common.util.StringUtils;
 import com.gamemoonchul.domain.entity.Member;
 import com.gamemoonchul.domain.entity.Post;
+import com.gamemoonchul.domain.entity.redis.RedisPostDetail;
 import com.gamemoonchul.infrastructure.web.dto.request.PostUploadRequest;
+import com.gamemoonchul.infrastructure.web.dto.response.MatchGameResponse;
+import com.gamemoonchul.infrastructure.web.dto.response.VoteRatioResponse;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Optional;
 
 public class PostConverter {
     public static Post requestToEntity(PostUploadRequest request, Member member) {
@@ -16,4 +24,56 @@ public class PostConverter {
             .build();
         return entity;
     }
+
+    public static RedisPostDetail toCache(Post post) {
+        return RedisPostDetail.builder()
+            .id(post.getId())
+            .author(MemberConverter.toResponseDto(post.getMember()))
+            .videoUrl(post.getVideoUrl())
+            .thumbnailUrl(post.getThumbnailUrl())
+            .commentCount(post.getCommentCount())
+            .title(post.getTitle())
+            .content(post.getContent())
+            .timesAgo(StringUtils.getTimeAgo(post.getCreatedAt()))
+            .viewCount(post.getViewCount())
+            .voteDetail(getVoteDetail(post))
+            .build();
+    }
+
+    private static List<VoteRatioResponse> getVoteDetail(Post post) {
+        HashMap<Long, Double> voteRatioMap = new HashMap<>();
+        post.getVoteOptions()
+            .forEach(vo -> {
+                voteRatioMap.put(vo.getId(), 0.0);
+                Optional.ofNullable(vo.getVotes())
+                    .ifPresent(votes -> {
+                        votes.forEach(v -> voteRatioMap.put(v.getId(), 0.0));
+                    });
+            });
+
+        Double totalVoteCnt = voteRatioMap.values()
+            .stream()
+            .mapToDouble(Double::doubleValue)
+            .sum();
+
+        voteRatioMap.forEach((k, v) -> {
+            if (totalVoteCnt == 0) {
+                voteRatioMap.put(k, 0.0);
+            } else {
+                voteRatioMap.put(k, (v / totalVoteCnt) * 100);
+            }
+        });
+
+        List<VoteRatioResponse> result = post.getVoteOptions()
+            .stream()
+            .map(vo -> {
+                Double voteRatio = voteRatioMap.get(vo.getId());
+                MatchGameResponse.MatchUserResponse matchUserResponse = MatchGameResponse.MatchUserResponse.toResponseVoId(vo.getMatchUser(), vo.getId());
+                return new VoteRatioResponse(matchUserResponse, voteRatio);
+            })
+            .toList();
+
+        return result;
+    }
+
 }
